+++ src/components/Header.js
@@ -39,8 +39,9 @@ const whatsNew = `
 - Removed default data
 - Added message default style/type can not be removed
 - When statrting the program the active style/type is now the same as when it was last closed
-- message improvements
+- Message improvements
 - After duplicating a style/type it becomes the active style/type
+- Bug fix where app would not fully close and use a lot cpu cycles
 `
 export class Header extends Component {